refactor(authentication): Created an authentication module, prompt password check for all pages
