+++ Homepage.py
@@ -1,36 +1,9 @@
 import streamlit as st
-import hmac
-import os
-
-st.set_page_config(layout="wide")
-
-def check_password():
-    """Returns `True` if the user had the correct password."""
-
-    def password_entered():
-        """Checks whether a password entered by the user is correct."""
-        password = os.getenv("STREAMLIT_PASSWORD", "")
-        if hmac.compare_digest(st.session_state["password"], password):
-            st.session_state["password_correct"] = True
-            del st.session_state["password"]  # Don't store the password.
-        else:
-            st.session_state["password_correct"] = False
-
-    # Return True if the password is validated.
-    if st.session_state.get("password_correct", False):
-        return True
-
-    # Show input for password.
-    st.text_input(
-        "Password", type="password", on_change=password_entered, key="password"
-    )
-    if "password_correct" in st.session_state:
-        st.error("😕 Password incorrect")
-    return False
+from authentication import check_password
 
 
 if not check_password():
-    st.stop()  # Do not continue if check_password is not True.
+    st.stop()
 
 import dashboard.database.processor_db as processor_db  # noqa: E402
 from dashboard import graph_functions as helper  # noqa: E402
