+++ authentication.py
@@ -1,0 +1,29 @@
+import streamlit as st
+import hmac
+import os
+
+st.set_page_config(layout="wide")
+
+def check_password():
+    """Returns `True` if the user had the correct password."""
+    def password_entered():
+        """Checks whether a password entered by the user is correct."""
+        env_password = os.getenv("STREAMLIT_PASSWORD", "")
+        if hmac.compare_digest(st.session_state["password"], env_password):
+            st.session_state["password_correct"] = True
+            del st.session_state["password"]  # Don't store the password.
+        else:
+            st.session_state["password_correct"] = False
+
+
+    # Return True if the password is validated.
+    if st.session_state.get("password_correct", False):
+        return True
+
+    # Show input for password.
+    st.text_input(
+        "Password", type="password", on_change=password_entered, key="password"
+    )
+    if "password_correct" in st.session_state:
+        st.error("😕 Password incorrect")
+    return False
